% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{add_hierarchy_level}
\alias{add_hierarchy_level}
\title{Add hierarchy level to self-referential table}
\usage{
add_hierarchy_level(hierarchy)
}
\arguments{
\item{hierarchy}{Data frame containing hierarchy data}
}
\description{
Add hierarchy level to self-referential table
}

