% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.bylevel}
\alias{aggregate.bylevel}
\title{Aggregate at level}
\usage{
\method{aggregate}{bylevel}(data, hierarchy, level)
}
\description{
Aggregate data by calling the top-level hierarchy that is required plus any
additional columns to be included.
}
\details{
Note that a hierarchy data frame is required, which is a data frame consisting
of the Class ID, Class Name and Parent ID.
}

