% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_atlevel}
\alias{aggregate_atlevel}
\title{Aggregate at level}
\usage{
aggregate_atlevel(data, hierarchy, level)
}
\arguments{
\item{data}{Data object to be included}

\item{hierarchy}{Data frame containing hierarchy data}

\item{level}{The level of interest}
}
\description{
Aggregate data at the requested level. The requested level will be the
cut-off point so that any levels below will be rolled up to the requested
level, and any levels above will remain as they are.
}
\details{
Note that a hierarchy data frame is required, which is a data frame consisting
of the Class ID, Class Name and Parent ID.
}

