% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_byname}
\alias{aggregate_byname}
\title{Aggregate data by concept name}
\usage{
aggregate_byname(data, hierarchy, concept)
}
\arguments{
\item{data}{Data object to be included}

\item{hierarchy}{Data frame containing hierarchy data}

\item{concept}{The concept name (column name) that needs to be aggregated}
}
\description{
Aggregate data by calling the top-level hierarchy that is required plus any
additional columns to be included.
}
\details{
Note that a hierarchy data frame is required, which is a data frame
consisting of the Class ID, Class Name and Parent ID.
}

