% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_list}
\alias{create_list}
\title{Create hierarchy as list of lists}
\usage{
create_list(hierarchy, level = 1)
}
\arguments{
\item{hierarchy}{Data frame containing hierarchy data}

\item{level}{The level of interest}
}
\description{
To work with networkD3 we need the hierarchy as a list of lists following
JSON structure.  This function takes the hierachy table and converts it to
the required list of lists format.
}
\details{
Since we are working with sel-referential tables, the depth of each
concept in the hierarchy needs to be identified first.

It accepts a level in case a lower level structure is required for the
visualization
}

