% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{get_children}
\alias{get_children}
\title{get_children by name}
\usage{
get_children(hierarchy, parent)
}
\arguments{
\item{hierarchy}{Data frame containing hierarchy data}

\item{data}{Data object to be included}
}
\description{
Returns a data.frame with all children of a named concept.
}

