% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect.concept}
\alias{inspect.concept}
\title{Concept Characteristics}
\usage{
inspect.concept(data, hierarchy)
}
\description{
Aggregate all data to the level given by (level 1 for the highest level). The
resulting data frame has column names as the aggregated columns.
}
\details{
It needs to answer: are all classes filled / balanced down to the lowest hierarchical level?
}

