% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect_concept}
\alias{inspect_concept}
\title{Concept Characteristics}
\usage{
inspect_concept(data, hierarchy, concept)
}
\arguments{
\item{data}{Data object to be included}

\item{hierarchy}{Data frame containing hierarchy data}

\item{concept}{The concept name (column name) that needs to be aggregated}
}
\description{
Aggregate all data to the level given by (level 1 for the highest level). The
resulting data frame has column names as the aggregated columns.
}
\details{
It needs to answer: are all classes filled / balanced down to the lowest
hierarchical level?
}

