% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect_hierarchy}
\alias{inspect_hierarchy}
\title{Hierarchy Characteristics}
\usage{
inspect_hierarchy(data, hierarchy)
}
\arguments{
\item{data}{Data object to be included}

\item{hierarchy}{Data frame containing hierarchy data}
}
\description{
When data gathering takes place for concept hierarchies there are two
approaches: Counts are made to the lowest possible level (this is common in
biological sciences), or counts are made to a specific level (as may be more
common in competency data, where the value is part of the definition of the
model).
}
\details{
We need to know which is which, so we test when a higher order concept has
children in the dataset whether these are already a sum of the lower
concepts or a different value. The result is a data frame with two columns:
`concept name` and `accumulated`. The latter has a yes/no value.
resulting data frame has column names as the aggregated columns.
}

