% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{jsonify}
\alias{jsonify}
\title{Create json formatted hierarchy}
\usage{
jsonify(hierarchy_as_list)
}
\arguments{
\item{hierarchy_as_list}{Hierarchy as a list of lists}
}
\description{
Small convenience function to convert the list of list as required by networkD3
to JSON format.
}

