% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/select.R
\name{select_level}
\alias{select_level}
\title{Select at level}
\usage{
select_level(data, hierarchy, level)
}
\description{
Get all the variables at the given level (level 1 for the highest level). If
the requested levels do not exist give a warning
}

