% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.bylevel}
\alias{visualize.bylevel}
\title{Visualize concept hierarchy by level}
\usage{
visualize.bylevel(data, hierarchy, level)
}
\arguments{
\item{data}{}

\item{hierarchy}{}

\item{level}{}
}
\description{
Visualize all the variables at the given level (level 1 for the highest
level) in a network3D diagram. If the requested levels do not exist give a
warning
}

