% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.ternary}
\alias{visualize.ternary}
\title{Visualize concept relationships by name}
\usage{
visualize.ternary(plot_categories, color)
}
\arguments{
\item{data}{}

\item{hierarchy}{}

\item{names}{vector of three names to include in ternary diagram}
}
\description{
Get all the variables at the given level (level 1 for the highest level). If
the requested levels do not exist give a warning
}

