% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_byname}
\alias{visualize_byname}
\title{Visualize concept hierarchy by name}
\usage{
visualize_byname(data, hierarchy, name)
}
\arguments{
\item{data}{}

\item{hierarchy}{}

\item{name}{name of highest level concept to visualize}
}
\description{
Get all the variables at the given level (level 1 for the highest level). If
the requested levels do not exist give a warning
}

