% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_hierarchy}
\alias{visualize_hierarchy}
\title{Visualize concept hierarchy}
\usage{
visualize_hierarchy(hierarchy)
}
\description{
Visualize all the variables at the given level (level 1 for the highest
level) in a network3D diagram. If the requested levels do not exist give a
warning
}

