% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_ternary}
\alias{visualize_ternary}
\title{Visualize concept relationships by name}
\usage{
visualize_ternary(data, hierarchy, plot_categories, color)
}
\description{
Get all the variables at the given level (level 1 for the highest level). If
the requested levels do not exist give a warning
}

