% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_level}
\alias{select_level}
\title{Select at level}
\usage{
select_level(data, hierarchy, level)
}
\arguments{
\item{data}{Data object to be included}

\item{hierarchy}{The hierachy definition}

\item{level}{The level of interest}
}
\description{
Get all the variables at the given level (level 1 for the highest level). If
the requested level does not exist give a warning.
}
\details{
Note that you are only selecting, not aggregating. If you want to roll-up
to any given level, use the aggregate_atlevel function instead.
}

