% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.ternary}
\alias{visualize.ternary}
\title{Visualize concept relationships by name}
\usage{
visualize.ternary(data, hierarchy, plot_categories, color)
}
\arguments{
\item{data}{}

\item{hierarchy}{}

\item{plot_categories}{}

\item{color}{}
}
\description{
Get all the variables at the given level (level 1 for the highest level). If
the requested levels do not exist give a warning
}

